+++ src/static/router.ts
@@ -4,29 +4,11 @@ export const router = createRouter({
   history: createWebHashHistory(),
   routes: [
     { path: '/', redirect: '/persons' },
-    {
-      path: '/persons',
-      component: () => import('./views/PersonsListView.vue'),
-    },
-    {
-      path: '/persons/:id',
-      component: () => import('./views/PersonDetailView.vue'),
-    },
-    {
-      path: '/places',
-      component: () => import('./views/PlacesListView.vue'),
-    },
-    {
-      path: '/places/:id',
-      component: () => import('./views/PlaceDetailView.vue'),
-    },
-    {
-      path: '/media',
-      component: () => import('../renderer/views/MediaView.vue'),
-    },
-    {
-      path: '/search',
-      component: () => import('../renderer/views/SearchView.vue'),
-    },
+    { path: '/persons', component: () => import('../renderer/views/PersonsView.vue') },
+    { path: '/persons/:personId', component: () => import('../renderer/views/PersonsView.vue') },
+    { path: '/places', component: () => import('../renderer/views/PlacesView.vue') },
+    { path: '/places/:id', component: () => import('../renderer/views/PlacesView.vue') },
+    { path: '/media', component: () => import('../renderer/views/MediaView.vue') },
+    { path: '/search', component: () => import('../renderer/views/SearchView.vue') },
   ],
 });